FRENDE/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.41
alembic==1.12.1
fastapi-users[sqlalchemy]==12.1.3
fastapi-users-db-sqlalchemy==7.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
pydantic-settings==2.2.1
aiosqlite==0.19.0
asyncpg==0.29.0
email-validator==2.1.2
websockets==11.0.3
PyJWT==2.8.0
psutil==5.9.6
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0
google-generativeai==0.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage[toml]==7.3.2
pytest-mock==3.12.0
# Database security and monitoring dependencies
cryptography==41.0.7
bcrypt==4.1.2
# Additional monitoring and security
prometheus-client==0.19.0
structlog==23.2.0
# Sentry error tracking
sentry-sdk[fastapi]==1.40.4
sentry-sdk[asgi]==1.40.4
redis==5.0.1
brotli==1.1.0
requests==2.31.0
boto3==1.34.0
psycopg2-binary==2.9.9
schedule==1.2.0
grafana-api==1.0.3
# Push notification dependencies
pywebpush==1.14.1 